+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+* Replace former `@cucumber/create-meta` with newer `@cucumber/ci-environment`
+
 ### Deprecated
 
 ### Removed